requirements.txt
langchain==0.2.1
langchain_openai==0.1.8
python-dotenv==1.0.1
streamlit==1.35.0





